[Notifications] Position popups relative to notification icon
The notification popup is horizontally centered around the notification icon so that when you have
a centered panel or the notification icon in a different place on the panel, the notification popup
are visually "emitted" by that icon. In System Tray, the set of popups is centered around the
System Tray applet.

For a vertical panel, it changes the position from bottom to top based on which half of the screen
the notification popup is on. Since notifications stack vertically, I found it odd to have them start
"in the middle of the screen", so I didn't implement centering vertically.

This affects only the default "near notification icon" placement setting. Manually setting a screen
corner still moves the notification popup all the way to the screen corner.